fix: resolve test races and stale assertion for issue #86
- Fix persisted_skills_survive_restart assertion: skills loaded from
  persist_dir have SkillLocation::User (not System) so they can
  override same-named builtins after restart (per dd0aec1)
- Move HOME_LOCK and HomeGuard from http::startup_tests to
  test_helpers so all tests that read HOME coordinate under the
  same mutex, eliminating the race between persisted_skills_survive_restart
  (which mutates HOME) and rule_check tests (which validate HOME)
