+++ crates/harness-server/src/test_helpers.rs
@@ -9,10 +9,39 @@ use harness_agents::AgentRegistry;
 use harness_core::HarnessConfig;
 
 /// Serialises every test that reads or mutates the process-global `HOME` env var.
-/// `tokio::test` runs tests concurrently in the same process; tests that call
-/// `tempdir_in_home` and then `validate_project_root` must hold this lock for
-/// the full duration to prevent a sibling test from mutating HOME in between.
+/// Tests that create tempdirs under HOME *and* tests that mutate HOME must both
+/// hold this lock to prevent races where one test mutates HOME while another
+/// test's `validate_project_root` reads it.
 pub static HOME_LOCK: tokio::sync::Mutex<()> = tokio::sync::Mutex::const_new(());
+
+/// RAII guard that restores `HOME` on drop, including on panic.
+pub struct HomeGuard {
+    original: Option<String>,
+}
+
+impl HomeGuard {
+    /// Overwrite `HOME` with `path` and return a guard that will undo the
+    /// change when dropped.
+    ///
+    /// # Safety
+    /// The caller must hold `HOME_LOCK` for the lifetime of this guard.
+    pub unsafe fn set(path: &std::path::Path) -> Self {
+        let original = std::env::var("HOME").ok();
+        std::env::set_var("HOME", path);
+        HomeGuard { original }
+    }
+}
+
+impl Drop for HomeGuard {
+    fn drop(&mut self) {
+        unsafe {
+            match self.original.take() {
+                Some(h) => std::env::set_var("HOME", h),
+                None => std::env::remove_var("HOME"),
+            }
+        }
+    }
+}
 
 /// Create a temp directory under a writable base path without mutating
 /// global state (`HOME` env var).  Tries `$HOME` first; falls back to
